refactor: errorCode 타입 변경 - Long -> Integer

+++ src/main/java/com/cjw/reactivecommunityproject/common/exception/model/RcBaseErrorMessage.java
@@ -3,7 +3,7 @@ package com.cjw.reactivecommunityproject.common.exception.model;
 import org.springframework.http.HttpStatus;
 
 public interface RcBaseErrorMessage {
-    Long getErrorCode();
+    Integer getErrorCode();
     String getErrorMessage();
     HttpStatus getHttpStatus();
     Boolean isDisplay();
