+++ src/main/java/com/cjw/reactivecommunityproject/web/auth/exception/AuthRestException.java
@@ -14,7 +14,7 @@ public class AuthRestException extends RcBaseException {
     }
 
     @Override
-    public Long getErrorCode() {
+    public Integer getErrorCode() {
         return super.errorCode;
     }
